+++ backend/core/js/tiny_mce/plugins/dextrose_videoembed/langs/fr.js
@@ -1,7 +1,0 @@
-tinyMCE.addI18n("fr.dextrose_videoembed",{
-	delta_width : 0,
-	delta_height: 0,
-	video: 'Ajouter une vidéo',
-	title: 'Ajouter une vidéo',
-	description: 'Collez votre code vidéo'
-});