+++ backend/core/js/tiny_mce/plugins/dextrose_videoembed/langs/nl.js
@@ -1,7 +1,0 @@
-tinyMCE.addI18n("nl.dextrose_videoembed",{
-	delta_width : 0,
-	delta_height: 0,
-	video: 'Video invoegen',
-	title: 'Video invoegen',
-	description: 'Plak hier uw code van uw video'
-});