+++ backend/core/js/tiny_mce/plugins/dextrose_videoembed/langs/ru.js
@@ -1,7 +1,0 @@
-tinyMCE.addI18n("ru.dextrose_videoembed",{
-	delta_width : 0,
-	delta_height: 0,
-	video: 'Добавить видео',
-	title: 'Добавить видео',
-	description: 'Вставить видео код'
-});