+++ frontend/core/engine/template.php
@@ -439,7 +439,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Format a number as currency
-	 * 	syntax: {$var|formatcurrency[:<currency>][:<decimals>]}
+	 * 	syntax: {$var|formatcurrency[:currency[:decimals]]}
 	 *
 	 * @param string $var The string to form.
 	 * @param string[optional] $currency The currency to will be used to format the number.
@@ -462,7 +462,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Format a number as a float
-	 * @later	grab settings from database
+	 * 	syntax: {$var|formatfloat[:decimals]}
 	 *
 	 * @param float $number The number to format.
 	 * @param int[optional] $decimals The number of decimals.
@@ -503,7 +503,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Get the navigation html
-	 * 	syntax: {$var|getnavigation[:<type>][:<parentId>][:<depth>][:<excludeIds-splitted-by-dash>]}
+	 * 	syntax: {$var|getnavigation[:type[:parentId[:depth[:excludeIds-splitted-by-dash]]]]}
 	 *
 	 * @param string[optional] $var The variable.
 	 * @param string[optional] $type The type of navigation, possible values are: page, footer.
@@ -529,7 +529,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Get a given field for a page-record
-	 * 	syntax: {$var|getpageinfo:404:'title'}
+	 * 	syntax: {$var|getpageinfo:pageId[:field[:language]]}
 	 *
 	 * @param string[optional] $var The string passed from the template.
 	 * @param int $pageId The id of the page to build the URL for.
@@ -558,6 +558,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Fetch the path for an include (theme file if available, core file otherwise)
+	 * 	syntax: {$var|getpath:file}
 	 *
 	 * @param string $var The variable.
 	 * @param string $file The base path.
@@ -573,7 +574,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Get the subnavigation html
-	 * 	syntax: {$var|getsubnavigation[:<type>][:<parentId>][:<startdepth>][:<enddepth>][:'<excludeIds-splitted-by-dash>']}
+	 * 	syntax: {$var|getsubnavigation[:type[:parentId[:startdepth[:enddepth[:'excludeIds-splitted-by-dash']]]]]}
 	 *
 	 * 	NOTE: When supplying more than 1 ID to exclude, the single quotes around the dash-separated list are mandatory.
 	 *
@@ -629,7 +630,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Get the URL for a given pageId & language
-	 * 	syntax: {$var|geturl:404}
+	 * 	syntax: {$var|geturl:pageId[:language]}
 	 *
 	 * @param string $var The string passed from the template.
 	 * @param int $pageId The id of the page to build the URL for.
@@ -649,7 +650,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Get the URL for a give module & action combination
-	 * 	syntax: {$var|geturlforblock:<module>:<action>:<language>}
+	 * 	syntax: {$var|geturlforblock:module[:action[:language]]}
 	 *
 	 * @param string $var The string passed from the template.
 	 * @param string $module The module wherefor the URL should be build.
@@ -671,6 +672,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Fetch an URL based on an extraId
+	 * 	syntax: {$var|geturlforextraid:extraId[:language]}
 	 *
 	 * @param string $var The string passed from the template.
 	 * @param int $extraId The id of the extra.
@@ -689,6 +691,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Highlights all strings in <code> tags.
+	 * 	syntax: {$var|highlight}
 	 *
 	 * @param string $var The string passed from the template.
 	 * @return string
@@ -717,6 +720,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Get a random var between a min and max
+	 * 	syntax: {$var|rand:min:max}
 	 *
 	 * @param string[optional] $var The string passed from the template.
 	 * @param int $min The miminum random number.
@@ -746,7 +750,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Formats a timestamp as a string that indicates the time ago
-	 * 	syntax: {$var|timeAgo}
+	 * 	syntax: {$var|timeago}
 	 *
 	 * @param string[optional] $var A UNIX-timestamp that will be formated as a time-ago-string.
 	 * @return string
@@ -764,7 +768,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Truncate a string
-	 * 	syntax: {$var|truncate:<max-length>[:<append-hellip>]}
+	 * 	syntax: {$var|truncate:max-length[:append-hellip]}
 	 *
 	 * @param string[optional] $var The string passed from the template.
 	 * @param int $length The maximum length of the truncated string.
@@ -801,7 +805,7 @@ class FrontendTemplateModifiers
 
 	/**
 	 * Get the value for a user-setting
-	 * 	syntax {$var|usersetting:<setting>[:<userId>]}
+	 * 	syntax {$var|usersetting:setting[:userId]}
 	 *
 	 * @param string[optional] $var The string passed from the template.
 	 * @param string $setting The name of the setting you want.