+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-3.1.10 (xxxx-xx-xx)
+3.2.0 (xxxx-xx-xx)
 --
+Improvements:
+
+* Core: integrated CKEditor into Fork CMS.
+* Pages: when adding an editor field, the editor will immediately open.
+
 Bugfixes:
 
 * Extensions: fixed module-warnings system.